// src/app/page/master/hari-form/hari-form.interface.ts
import {EventEmitter} from '@angular/core';

export interface Hari {
    kodeHari?: string;
	namaHari?: string;
	formatAngka?: number;
	formatAngkaRomawi?: string;
	noUrutHariKe?: number;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	id?: any;
	kdProfile?: string;
	kode?: any;
	version?: any;
	label?: any;
}